refactor(organise): better group app code
Put all the app code into a `./src` directory to make it more obvious what the
actual app is, and make those files easier to target.
Update `package.json` and `snowtify` to reflect the refactor, so that there are
NO breaking changes.
