+++ test/logger-mock/log-tester.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-const logger = require('../../logger');
+const logger = require('../../src/logger');
 const logs   = require('./logs.json');
 
 Object.keys(logs).forEach(level => logger.log(level, logs[level]));
